requirements.txt
django
djangorestframework
django-cors-headers
django-environ
djangorestframework-simplejwt
django-filter
mysqlclient
